// Atomics Manager/ClientApp/src/app/admin/models/bon-livraison.ts
export class BonLivraison {
   id: number;
   refBL: string;
   demandesName: string;
   demandesId: number;
   livreure: string;
   dateLivraison: Date;
   isInSla: boolean;
   control: boolean;
   controleurId: string;
   controleurUserName: string;
   matchToBon: boolean;
   commentaire: string;
   noteFournisseur: number;
}